requirements.txt
aiogram
asyncpg
sqlalchemy
python-dotenv
